requirements.txt
flask==2.0.3
werkzeug==2.0.3
pandas==1.5.3
python-dotenv==0.21.1
google-cloud-dialogflow==2.16.0
gunicorn==20.1.0
google-cloud-translate==3.8.0